+++ setuptools_behave.py
@@ -6,6 +6,9 @@ Setuptools command for behave.
 .. code-block:: console
 
     python setup.py behave_test
+    python setup.py behave_test --format=progress3
+    python setup.py behave_test --args=features/one.feature
+    python setup.py behave_test --tags=-xfail --args=features
 
 .. seealso::
 
@@ -113,7 +116,9 @@ class behave_test(Command):
     def behave(self, path):
         behave = os.path.join("bin", "behave")
         if not os.path.exists(behave):
-            behave = "behave"
+            # -- ALTERNATIVE: USE: behave script: behave = "behave"
+            # -- USE: behave module (main)
+            behave = "-m behave"
         cmd_options = ""
         if self.tags:
             cmd_options = "--tags=" + " --tags=".join(self.tags)