+++ CHANGES.rst
@@ -21,6 +21,7 @@ ENHANCEMENTS:
 
 FIXED:
 
+  * issue #310: Use setuptools_behave.py with behave module
   * issue #309: behave --lang-list fails on Python3 (and Python2)